refactor(lsp): dedupe uri/pointer helpers, rename shadowed type, add ADR 01067
Code-review cleanups on the phase 3 diff:
- Rename the model's `SyntaxError` interface to `SyntaxErrorSpan` so it no longer
  shadows the global built-in `SyntaxError`.
- Move isJsonUri/isYamlUri to gate.ts (with the other extension logic) and use
  them in completion.ts and hover.ts instead of inline `.endsWith(".json")`.
- Export pointerFromPath from json/positions and reuse it in yaml/positions,
  removing the duplicated JSON-Pointer escaping.
- Add ADR 01067 recording the YAML-parity SpecModel and the warn-not-error
  decision for valid v2 specs (behavior change; per-phase ADR per CLAUDE.md).

No behavior change; dist/lsp stays at 100% coverage (81 lsp tests).
